move proto out of internal
