+++ elasticsearch.go
@@ -2,7 +2,7 @@ package goreplay
 
 import (
 	"encoding/json"
-	"github.com/buger/goreplay/internal/proto"
+	"github.com/buger/goreplay/proto"
 	"log"
 	"net/url"
 	"strings"
