+++ input_raw.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"fmt"
 	"github.com/buger/goreplay/internal/capture"
-	"github.com/buger/goreplay/internal/proto"
 	"github.com/buger/goreplay/internal/tcp"
+	"github.com/buger/goreplay/proto"
 	"log"
 	"net"
 	"strconv"
