+++ internal/capture/capture.go
@@ -5,9 +5,9 @@ import (
 	"errors"
 	"expvar"
 	"fmt"
-	"github.com/buger/goreplay/internal/proto"
 	"github.com/buger/goreplay/internal/size"
 	"github.com/buger/goreplay/internal/tcp"
+	"github.com/buger/goreplay/proto"
 	"io"
 	"log"
 	"net"
