+++ proto/fuzz.go
